requirements.txt

django-ckeditor==5.6.1
django-js-asset==1.1.0
Pillow==5.3.0
psycopg2==2.7.5
pytz==2018.6